requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
sentence-transformers
scikit-learn
python-docx
streamlit
requests
google-generativeai
python-multipart
datetime
loguru